requirements.txt
# Dépendances principales pour l'application Flask avec ChatGPT-4
flask==2.3.3
requests==2.31.0
openai==1.3.7
python-dotenv==1.0.0

# Dépendances pour le serveur SAP PM
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2

# Dépendances pour SAP PM (base de données en mémoire)
# Note: MongoDB a été remplacé par une base de données en mémoire
# motor==3.3.2
# pymongo==4.6.0 